fix(api): Implements "Remember Me" functionality
Adds the ability to extend the user session duration.
The login endpoint now accepts an optional "extended" boolean
parameter. If true, the generated JWT will have a longer
expiration time (30 days). Otherwise, the default expiration
time (7 days) will be used. This is enabled via the extended
parameter in the login request and updates the related schema,
route, and test files.